requirements.txt
aiohttp
anitopy
apscheduler
aria2p
country_converter
emoji-country-flag
html_telegraph_poster
telethon
tgcrypto
psutil
pycountry
pymediainfo
pymongo[srv]
pyrogram==2.0.106
#pyrogram==2.0.73
python-decouple
qbittorrent-api
random_word
tzlocal
#Depreciating...
quote
